==> netherlands_amsterdam/030-training_models/5.creating_extended_df.ipynb <==
import pandas as pd
import sklearn
import matplotlib as plt
from pycaret.regression import *

df = pd.read_csv('1.complete_df.csv')
df.set_index('Unnamed: 0', inplace = True)
df

cardinal_directions = ["N", "S", "E", "W"]
diagonal_directions = ["NW", "NE", "SW", "SE"]

categories = ['nature', 'entertainment', 'transports', 'art', 'sport', 'tourism', 'security', 'accessibility']

def add_cardinal_features(source_df, categories_names, direction: str):

    new_categories_names = []
    df = source_df.copy()
    df[['row', 'column']] = df[['row', 'column']].astype(int)

    if direction == 'N':
        for category_name in categories_names:
            new_categories_names.append(category_name + '_north')

        df = df.loc[df.row > 1, ['row', 'column'] + categories_names]
        df['row'] = df['row'] - 1

    if direction == 'S':
        for category_name in categories_names:
            new_categories_names.append(category_name + '_south')

        df = df.loc[df.row < df.row.max(), ['row', 'column'] + categories_names]
        df['row'] = df['row'] + 1

    if direction == 'W':
        for category_name in categories_names:
           new_categories_names.append(category_name + '_west')

        df = df.loc[df.column > 1, ['row', 'column'] + categories_names]
        df['column'] = df['column'] - 1

    if direction == 'E':
        for category_name in categories_names:
            new_categories_names.append(category_name + '_east')

        df = df.loc[df.column < df.column.max(), ['row', 'column'] + categories]
        df['column'] = df['column'] + 1

    df[['row', 'column']] = df[['row', 'column']].astype(str)
    df.columns = ['row', 'column'] + new_categories_names
    df.index = 'R' + df.row + 'C' + df.column

    df = df[new_categories_names]

    return df

def add_diagonal_features(source_df, categories_names, direction: str):

    new_categories_names = []
    df = source_df.copy()
    df[['row', 'column']] = df[['row', 'column']].astype(int)

    if direction == 'NW':
        for category_name in categories_names:
            new_categories_names.append(category_name + '_north_west')

        df = df.loc[(df.row > 1) & (df.column > 1), ['row', 'column'] + categories_names]
        df['row'] = df['row'] - 1
        df['column'] = df['column'] - 1

    if direction == 'SW':
        for category_name in categories_names:
            new_categories_names.append(category_name + '_south_west')

        df = df.loc[(df.row < df.row.max()) & (df.column > 1), ['row', 'column'] + categories_names]
        df['row'] = df['row'] + 1
        df['column'] = df['column'] - 1

    if direction == 'NE':
        for category_name in categories_names:
            new_categories_names.append(category_name + '_north_east')

        df = df.loc[(df.row > 1) & (df.column < df.column.max()), ['row', 'column'] + categories_names]
        df['row'] = df['row'] - 1
        df['column'] = df['column'] + 1

    if direction == 'SE':
        for category_name in categories_names:
            new_categories_names.append(category_name + '_south_east')

        df = df.loc[(df.row < df.row.max()) & (df.column < df.column.max()), ['row', 'column'] + categories]
        df['row'] = df['row'] + 1
        df['column'] = df['column'] + 1

    df[['row', 'column']] = df[['row', 'column']].astype(str)
    df.columns = ['row', 'column'] + new_categories_names
    df.index = 'R' + df.row + 'C' + df.column

    df = df[new_categories_names]

    return df

extended_df = df.copy()

new_df = add_cardinal_features(df, categories, direction = 'W')
extended_df = pd.concat([extended_df, new_df], axis = 1)

new_df = add_cardinal_features(df, categories, direction = 'E')
extended_df = pd.concat([extended_df, new_df], axis = 1)

new_df = add_cardinal_features(df, categories, direction = 'S')
extended_df = pd.concat([extended_df, new_df], axis = 1)

new_df = add_cardinal_features(df, categories, direction = 'W')
extended_df = pd.concat([extended_df, new_df], axis = 1)

extended_df

new_df = add_diagonal_features(df, categories, direction = 'NE')
extended_df = pd.concat([extended_df, new_df], axis = 1)

new_df = add_diagonal_features(df, categories, direction = 'SE')
extended_df = pd.concat([extended_df, new_df], axis = 1)

new_df = add_diagonal_features(df, categories, direction = 'SW')
extended_df = pd.concat([extended_df, new_df], axis = 1)

new_df = add_diagonal_features(df, categories, direction = 'NW')
extended_df = pd.concat([extended_df, new_df], axis = 1)

extended_df

extended_df.fillna(value = 0, inplace = True)
extended_df

extended_df.to_csv('5.extended_df.csv')

categories
